requirements.txt
Flask==3.0.3
flask-cors==5.0.0
gunicorn==23.0.0
Jinja2==3.1.4
openai-whisper==20231117
# Install PyTorch with CUDA support
torch==2.3.1+cu121 --index-url https://download.pytorch.org/whl/cu121


